Use stdlib inspect.signature if available.
cf #8822.
